+++ lib/src/cook/chai/module/ToolchainElement.cpp
@@ -16,6 +16,7 @@ namespace cook { namespace chai { namespace module {
 
         EXPOSE_VALUE(Part, Cli);
         EXPOSE_VALUE(Part, Pre);
+        EXPOSE_VALUE(Part, Runtime);
         EXPOSE_VALUE(Part, Deps);
         EXPOSE_VALUE(Part, Export);
         EXPOSE_VALUE(Part, Output);
@@ -29,6 +30,7 @@ namespace cook { namespace chai { namespace module {
         EXPOSE_VALUE(Part, LibraryPath);
         EXPOSE_VALUE(Part, Framework);
         EXPOSE_VALUE(Part, FrameworkPath);
+        EXPOSE_VALUE(Part, Resource);
 
         EXPOSE_TYPE(ElementType);
         EXPOSE_VALUE(ElementType, Compile);