+++ lib/src/cook/process/toolchain/serialize/MSVC.cpp
@@ -16,6 +16,7 @@ namespace cook { namespace process { namespace toolchain { namespace serialize {
         if (false) {
         } else if (k == "debug_symbols" && v == "true") {
             kv.append(Part.Pre, "Zi")
+            tm[Part.Runtime] = fun(k,v) { return "/${k}d" }
         } else if (k == "optimization" && v == "max_speed") {
             kv.append(Part.Pre, "O2")
         } else if (k == "arch" && v == "x86") {
@@ -23,8 +24,8 @@ namespace cook { namespace process { namespace toolchain { namespace serialize {
         } else if (k == "position_independent_code" && v == "true") {
         } else if (k == "c++.runtime") {
             if (false) {}
-            else if (v == "dynamic") { kv.append(Part.Pre, "MD") }
-            else if (v == "static") { kv.append(Part.Pre, "MT") }
+            else if (v == "dynamic") { kv.append(Part.Runtime, "MD") }
+            else if (v == "static") { kv.append(Part.Runtime, "MT") }
         } else if (k == "c++.std" && e.language == Language.CXX) {
             kv.append(Part.Pre, "std", "c++${v}")
         } else {
@@ -47,7 +48,7 @@ namespace cook { namespace process { namespace toolchain { namespace serialize {
 })
 
 
-for( lang : [Language.C, Language.CXX, Language.ASM]) {
+for( lang : [Language.C, Language.CXX, Language.ASM, Language.Resource]) {
 
     var compiler = cook.toolchain.element(ElementType.Compile, lang, TargetType.Object)
     var & kv = compiler.key_values()
@@ -55,22 +56,32 @@ for( lang : [Language.C, Language.CXX, Language.ASM]) {
 
     tm[Part.Cli]            = fun(k,v) { return k }
     tm[Part.Pre]            = fun(k,v) { if (v.empty) { return "/${k}" } else { return "/${k}:${v}" } }
+    tm[Part.Runtime]        = fun(k,v) { return "/${k}" }
     tm[Part.Output]         = fun(k,v) { return "/Fo${k}" }
     tm[Part.Input]          = fun(k,v) { return k }
     tm[Part.Deps]           = fun(k,v) { return "/showIncludes" }
     tm[Part.Define]         = fun(k,v) { if (v.empty) { return "/D${k}" } else { return "/D${k}=${v}" } }
     tm[Part.IncludePath]    = fun(k,v) { if (k.empty) { return "/I./" } else { return "/I${k}" } }
     tm[Part.ForceInclude]   = fun(k,v) { return "/FI${k}" }
+    tm[Part.Resource]       = fun(k,v) { return "" }
 
     if (lang == Language.C) {
         kv.append(Part.Pre, "TC")
     }
-    kv.append(Part.Cli, "cl")
-    kv.append(Part.Pre, "nologo")
+    if (lang == Language.Resource) {
+        kv.append(Part.Cli, "rc")
+        tm[Part.Pre]        = fun(k,v) { return "" }
+        tm[Part.Runtime]    = fun(k,v) { return "" }
+        tm[Part.Deps]       = fun(k,v) { return "" }
+        tm[Part.Resource]   = fun(k,v) { return k }
+    } else {
+        kv.append(Part.Cli, "cl")
+    }
     kv.append(Part.Pre, "EHsc")
     kv.append(Part.Pre, "bigobj")
     kv.append(Part.Pre, "FS")
     kv.append(Part.Pre, "c")
+    kv.append(Part.Pre, "nologo")
     kv.append(Part.Define, "NOMINMAX")
     kv.append(Part.Deps, "msvc")
 }
@@ -98,7 +109,7 @@ for(s : [TargetType.Executable, TargetType.SharedLibrary]){
     tm[Part.Output]         = fun(k,v) { return "/OUT:${k}" }
     tm[Part.Input]          = fun(k,v) { return k }
     tm[Part.Library]        = fun(k,v) { return "${k}.lib" }
-    tm[Part.LibraryPath]    = fun(k,v) { if (k.empty) { return "/LIBPATH:./" } else { return "/LIBPATH:${k}" } }
+    tm[Part.LibraryPath]    = fun(k,v) { if (k.empty) { return "/LIBPATH:./" } else { return "/LIBPATH:\"${k}\"" } }
     
     kv.append(Part.Cli, "link")
     if (s == TargetType.SharedLibrary){